Course_Projects/03_coffee_shop/backend/requirements.txt
astroid==3.0.2
Click==8.1.3
ecdsa==0.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
future==0.17.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.10.0
MarkupSafe==2.1.1
mccabe==0.7.0
pycryptodome==3.3.1
pylint==3.0.3
python-jose-cryptodome==1.3.2
python-dotenv==1.0.0
six==1.16.0
typed-ast==1.4.2
Werkzeug==3.0.1
wrapt==1.14.1
